+++ examples/017-frame_text_flow.rb
@@ -28,7 +28,7 @@ boxes = []
 boxes << doc.layout.image_box(File.join(__dir__, 'machupicchu.jpg'),
                               width: 100, margin: [10, 30], position: :float)
 boxes << Box.create(width: 50, height: 50, margin: 20,
-                    position: :float, position_hint: :right,
+                    position: :float, align: :right,
                     background_color: "hp-blue-light2",
                     border: {width: 1, color: "hp-blue-dark"})
 boxes << doc.layout.lorem_ipsum_box(count: 3, position: :flow, text_align: :justify)